backend/requirements.txt
# FastAPI and server
fastapi>=0.120.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.20

# Data validation
pydantic>=2.8.2
pydantic-settings>=2.6.1

# Environment and config
python-dotenv>=1.1.1

# HTTP requests
requests>=2.32.5
httpx>=0.26,<0.29
aiohttp>=3.13.1

# Fetch.ai uAgents (use latest compatible version)
uagents>=0.22.0
cosmpy>=0.9.2

# WebSockets
websockets>=15.0.1

# Web Scraping
playwright>=1.49.0

# Database
sqlalchemy>=2.0.36
alembic>=1.13.1
supabase>=2.22.2

# ngrok for webhooks
ngrok>=1.4.0

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Code quality
black>=25.1.0
ruff>=0.14.2
mypy>=1.15.0

# Document parsing
llama-parse>=0.5.17
google-generativeai>=0.8.5
